Allow test suite to handle stacks tools that write output to input dir.
We now allow the input directory to be used as an output directory as well.
Under the hood, we're simply making a copy of the input directory (so that we
may delete it later for cleanup).

To tell ok_() that it should use the input directory as the output
directory, pass it the -i flag with argument set to the input directory.
The command under test MUST use the %in test suite variable.  In the
special case that the inline input directory is %in, then just call
-i '%in'.

Two examples follow.


$freq_in='some_special_input_directory'

ok_ -i '%in' \
    'input fastq' \
    001_input_fastq \
    "process_radtags -i fastq -p %in -o %out -b $barcodes -E phred33 -e sbfI"

ok_ -i $freq_in \
    'input gzfastq' \
    000_input_gzfastq \
    "process_radtags -i gzfastq -f %in/in.fastq.gz -o %out -b $barcodes -E phred33 -e sbfI"