populations --vcf: removed support for GL field; removed the now unused member Datum::snps.
There could still be issues with how the SNPs LRTs interacted with CIGARs in the
"merged" loci. Also, having the GL field was slow, and its contents did not match
the standard.

It seemed more simple to put these issues aside until we add full support for GQ
and GL.

The code that was removed in `populations.cc:merge_datums()` was doing nothing
because it looped on `Datum::snps`, empty until write_vcf was called (just
before the end of `main()`).