Noting that 8bc847568931 changed the SNPs files of 'bounded' and 'fixed', making them consistent with those of 'snp'.
The SNPs files for model 'snp' haven't changed.

Previously the `if(no rank 2 observations)` check when creating ambiguous SNPs
was only applied in the 'snp' model. Thus the rank 2 nucleotide in the SNPs
files was defaulting to A (or C if the rank 1 nucleotide was A) when a single
nucleotide was observed but the call was unknown because of a low depth.