New program tsv2bam.
This program will be analogous to the former tags2bam & tags2bam_pe. I am
removing the writing BAM files at the end of sstacks, because (1) there's no
practical way to specify paired-end read files when working with several
samples, and (2) it will make for a clearer separation between the new
components of the pipeline and the conserved one.