tsv2bam: Hopefully solved the "forward reads alignment to the catalog" issue for good (c.f. c53d31d62da9, ed33399c8600).
The trimming that I orginially performed (and that the assert revealed) was the right idea.
It did not, however, handle variation in read length. Now we pad to the length of the
catalog rather than trimming. (N.B. Match files CIGARs do not describe this padding: a
match may have a CIGAR shorter than the catalog locus, or no cigar at all and still need
terminal padding.)